Implement TraineeDAO, TrainerDAO, TraningDAO CRUD methods

+++ sping_core_task/src/main/java/com/donec/spring_core_task/dao/TrainerDAO.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface TrainerDAO {
-    Optional<Trainer> create(int trainerId, Trainer trainer);
+    Optional<Trainer> create(long trainerId, Trainer trainer);
     Optional<Trainer> update (Trainer trainer);
-    Optional<Trainer> getById(int trainerId);
+    Optional<Trainer> getById(long trainerId);
     Optional<Trainer> getByUsername(String username);
     List<Trainer> getAllTrainers();
     List<Trainer> getAllTrainersByUsername(String username);
